+++ src/test/resources/db/formula/formula_insert_log_011001110100000_mysql.sql
@@ -6,7 +6,7 @@ CREATE PROCEDURE formula_insert_log_011001110100000
  _field_id_formula_name    smallint,
  _field_id_user_id         smallint,
  _field_id_formula_type_id smallint,
- _formula_type_id          bigint,
+ _formula_type_id          smallint,
  _field_id_formula_text    smallint,
  _formula_text             text,
  _field_id_resolved_text   smallint,
